backend/requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
Flask-Bcrypt==1.0.1
PyJWT==2.7.0
python-dotenv==1.0.0
